// types/weather-conditions.ts
export enum WeatherCondition {
  SKC = "skc",
  FEW = "few",
  SCT = "sct",
  BKN = "bkn",
  OVC = "ovc",
  WIND_SKC = "wind_skc",
  WIND_FEW = "wind_few",
  WIND_SCT = "wind_sct",
  WIND_BKN = "wind_bkn",
  WIND_OVC = "wind_ovc",
  SNOW = "snow",
  RAIN_SNOW = "rain_snow",
  RAIN_SLEET = "rain_sleet",
  SNOW_SLEET = "snow_sleet",
  FZRA = "fzra",
  RAIN_FZRA = "rain_fzra",
  SNOW_FZRA = "snow_fzra",
  SLEET = "sleet",
  RAIN = "rain",
  RAIN_SHOWERS = "rain_showers",
  RAIN_SHOWERS_HI = "rain_showers_hi",
  TSRA = "tsra",
  TSRA_SCT = "tsra_sct",
  TSRA_HI = "tsra_hi",
  TORNADO = "tornado",
  HURRICANE = "hurricane",
  TROPICAL_STORM = "tropical_storm",
  DUST = "dust",
  SMOKE = "smoke",
  HAZE = "haze",
  HOT = "hot",
  COLD = "cold",
  BLIZZARD = "blizzard",
  FOG = "fog",
}